feat: Split up form models into individual files
Tests as well. Should be otherwise equivalent
